Continuous Assessment-ONE for B9IS123-TMD1S
will appear at 09:00 on December 1

==============================================

Create a class venue, with attributes capacity, roomNumber, set in the constructor. 

Create a method addModule, assigning a module to a timetable of day/time listings for a college week.

20

Create a class module, with attributes subject, location and hours, and optionally day, startTime and lecturer, set in the constructor.

Create a method to set the lecturer, which will verify that the lecturer is available and teaches the subject, and returns whether the assignment was successful

Create a method to set the timetable for the subject, which will verify that the lecturer is available, and returns whether the assignment was successful

25

Create a class lecturer, with attributes name and location, set in the constructor. It should be possible to add subjects to their set of subjects taught.

15

Create a class student, with attributes name and course, set in the constructor

create a method addModule, assigning the student to a module. It should fail if the class is at capacity, or there is a timetabling conflict

20

Ensure invalid inputs and operations result in appropriate errors

20